fix: address extension followup ci failures
